requirements.txt
requests>=2.31.0
tabulate>=0.9.0
python-dateutil>=2.9.0
click>=8.0.0